Move to screaming architecture
