+++ tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/crew/model/Crew.kt
@@ -1,0 +1,15 @@
+package ru.thetenzou.tsoddservice.crew.model
+
+import javax.persistence.*
+
+@Entity
+@Table(name = "crews")
+data class Crew(
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    var id: Long,
+
+    @Column(name = "name")
+    var name: String,
+)
