--- tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/model/tsodd/TsoddName.kt
@@ -1,16 +1,0 @@
-package ru.thetenzou.tsoddservice.model.tsodd
-
-import javax.persistence.*
-
-@Entity
-@Table(name = "tsodd_names")
-data class TsoddName(
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    var id: Long,
-    @Column(name = "name")
-    var name: String,
-    @ManyToOne
-    @JoinColumn(name = "type_id")
-    var tsoddType: TsoddType,
-)
