--- tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/repository/task/TaskRepository.kt
@@ -1,8 +1,0 @@
-package ru.thetenzou.tsoddservice.repository.task
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.rest.core.annotation.RepositoryRestResource
-import ru.thetenzou.tsoddservice.model.task.Task
-
-@RepositoryRestResource
-interface TaskRepository : JpaRepository<Task, Long>
