--- tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/repository/tsodd/TsoddConditionRepository.kt
@@ -1,8 +1,0 @@
-package ru.thetenzou.tsoddservice.repository.tsodd
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.rest.core.annotation.RepositoryRestResource
-import ru.thetenzou.tsoddservice.model.tsodd.TsoddCondition
-
-@RepositoryRestResource
-interface TsoddConditionRepository : JpaRepository<TsoddCondition, Long>
