--- tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/repository/tsodd/TsoddTypeRepository.kt
@@ -1,8 +1,0 @@
-package ru.thetenzou.tsoddservice.repository.tsodd
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.rest.core.annotation.RepositoryRestResource
-import ru.thetenzou.tsoddservice.model.tsodd.TsoddType
-
-@RepositoryRestResource
-interface TsoddTypeRepository : JpaRepository<TsoddType, Long>
