--- tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/request/SolveRequestDto.kt
@@ -1,9 +1,0 @@
-package ru.thetenzou.tsoddservice.request
-
-import java.time.LocalDate
-
-data class SolveRequestDto(
-    val name: String,
-    val startDate: LocalDate,
-    val endDate: LocalDate,
-)
