+++ tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/schedule/repository/ScheduleRepository.kt
@@ -1,0 +1,8 @@
+package ru.thetenzou.tsoddservice.schedule.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.rest.core.annotation.RepositoryRestResource
+import ru.thetenzou.tsoddservice.schedule.model.Schedule
+
+@RepositoryRestResource
+interface ScheduleRepository : JpaRepository<Schedule, Long>
