+++ tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/schedule/repository/ScheduledTaskRepository.kt
@@ -1,0 +1,8 @@
+package ru.thetenzou.tsoddservice.schedule.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.rest.core.annotation.RepositoryRestResource
+import ru.thetenzou.tsoddservice.schedule.model.ScheduledTask
+
+@RepositoryRestResource
+interface ScheduledTaskRepository : JpaRepository<ScheduledTask, Long>
