+++ tsodd-service/src/main/kotlin/ru/thetenzou/tsoddservice/schedule/request/SolveRequestDto.kt
@@ -1,0 +1,9 @@
+package ru.thetenzou.tsoddservice.schedule.request
+
+import java.time.LocalDate
+
+data class SolveRequestDto(
+    val name: String,
+    val startDate: LocalDate,
+    val endDate: LocalDate,
+)
